// src/data/authors.generated.ts
export const authors: AuthorRecord[] = [
  {
    "authorId": "idb0000-alqubalee-abdullah",
    "openAlexId": "A5035537771",
    "openAlexIds": [
      "A5035537771"
    ],
    "name": "Abdullah Alqubalee",
    "groupId": "",
    "groupName": "",
    "affiliate1": "",
    "affiliate2": "",
    "affiliate3": "",
    "email": "abdullah.alqubalee@kfupm.edu.sa",
    "orcid": "0000-0002-0543-9026",
    "totalPublications": 45,
    "totalCitations": 390,
    "hIndex": 12
  },
// ...
];